[DB] renames to db
